Move things from JavaRuleUtil to JavaAstUtils
