+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/bestpractices/UnusedPrivateFieldRule.java
@@ -14,6 +14,7 @@ import net.sourceforge.pmd.lang.java.ast.ASTFieldDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTVariableDeclaratorId;
 import net.sourceforge.pmd.lang.java.ast.AccessNode.Visibility;
 import net.sourceforge.pmd.lang.java.ast.JavaNode;
+import net.sourceforge.pmd.lang.java.ast.internal.JavaAstUtils;
 import net.sourceforge.pmd.lang.java.rule.AbstractJavaRulechainRule;
 import net.sourceforge.pmd.lang.java.rule.internal.JavaPropertyUtil;
 import net.sourceforge.pmd.lang.java.rule.internal.JavaRuleUtil;
@@ -48,14 +49,14 @@ public class UnusedPrivateFieldRule extends AbstractJavaRulechainRule {
     public Object visitJavaNode(JavaNode node, Object data) {
         if (node instanceof ASTAnyTypeDeclaration) {
             ASTAnyTypeDeclaration type = (ASTAnyTypeDeclaration) node;
-            if (JavaRuleUtil.hasAnyAnnotation(type, INVALIDATING_CLASS_ANNOT)) {
+            if (JavaAstUtils.hasAnyAnnotation(type, INVALIDATING_CLASS_ANNOT)) {
                 return null;
             }
 
             for (ASTFieldDeclaration field : type.getDeclarations().filterIs(ASTFieldDeclaration.class)) {
                 if (!isIgnored(field)) {
                     for (ASTVariableDeclaratorId varId : field.getVarIds()) {
-                        if (JavaRuleUtil.isNeverUsed(varId)) {
+                        if (JavaAstUtils.isNeverUsed(varId)) {
                             addViolation(data, varId, varId.getName());
                         }
                     }
@@ -69,7 +70,7 @@ public class UnusedPrivateFieldRule extends AbstractJavaRulechainRule {
         return field.getVisibility() != Visibility.V_PRIVATE
             || JavaRuleUtil.isSerialPersistentFields(field)
             || JavaRuleUtil.isSerialVersionUID(field)
-            || JavaRuleUtil.hasAnyAnnotation(field, getProperty(IGNORED_FIELD_ANNOTATIONS));
+            || JavaAstUtils.hasAnyAnnotation(field, getProperty(IGNORED_FIELD_ANNOTATIONS));
     }
 
 }
