+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/errorprone/BrokenNullCheckRule.java
@@ -11,8 +11,8 @@ import net.sourceforge.pmd.lang.java.ast.ASTInfixExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTNullLiteral;
 import net.sourceforge.pmd.lang.java.ast.BinaryOp;
 import net.sourceforge.pmd.lang.java.ast.QualifiableExpression;
+import net.sourceforge.pmd.lang.java.ast.internal.JavaAstUtils;
 import net.sourceforge.pmd.lang.java.rule.AbstractJavaRulechainRule;
-import net.sourceforge.pmd.lang.java.rule.internal.JavaRuleUtil;
 import net.sourceforge.pmd.lang.java.rule.internal.StablePathMatcher;
 
 public class BrokenNullCheckRule extends AbstractJavaRulechainRule {
@@ -46,7 +46,7 @@ public class BrokenNullCheckRule extends AbstractJavaRulechainRule {
             return;
         }
 
-        ASTExpression otherChild = JavaRuleUtil.getOtherOperandIfInInfixExpr(nullLit);
+        ASTExpression otherChild = JavaAstUtils.getOtherOperandIfInInfixExpr(nullLit);
         StablePathMatcher pathToNullVar = StablePathMatcher.matching(otherChild);
         if (pathToNullVar == null) {
             // cannot be matched, because it's not stable
