+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/rule/internal/JavaRuleUtilTest.java
@@ -6,8 +6,6 @@ package net.sourceforge.pmd.lang.java.rule.internal;
 
 import static net.sourceforge.pmd.lang.java.rule.internal.JavaRuleUtil.containsCamelCaseWord;
 import static net.sourceforge.pmd.lang.java.rule.internal.JavaRuleUtil.startsWithCamelCaseWord;
-import static net.sourceforge.pmd.util.CollectionUtil.listOf;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -15,9 +13,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.Test;
 
 import net.sourceforge.pmd.lang.java.BaseParserTest;
-import net.sourceforge.pmd.lang.java.ast.ASTExpression;
-import net.sourceforge.pmd.lang.java.ast.ASTInfixExpression;
-import net.sourceforge.pmd.lang.java.ast.ASTVariableAccess;
 
 public class JavaRuleUtilTest extends BaseParserTest {
 
@@ -43,28 +38,6 @@ public class JavaRuleUtilTest extends BaseParserTest {
         assertThrows(NullPointerException.class, () -> containsCamelCaseWord("fnei", null));
         assertThrows(AssertionError.class, () -> containsCamelCaseWord("fnei", ""), "empty string");
         assertThrows(AssertionError.class, () -> containsCamelCaseWord("fnei", "a"), "not capitalized");
-    }
-
-    @Test
-    public void testFlattenConcatOperands() {
-        ASTExpression e = parseExpr("s1+s2+s3");
-
-        assertTrue(JavaRuleUtil.isStringConcatExpr(e));
-        assertEquals(e.descendants(ASTVariableAccess.class).toList(),
-                     JavaRuleUtil.flattenOperands(e).toList());
-    }
-
-    @Test
-    public void testFlattenConcatOperandsRespectsTyping() {
-        ASTInfixExpression e = (ASTInfixExpression) parseExpr("i+j+s2+s3");
-        assertTrue(JavaRuleUtil.isStringConcatExpr(e));
-        ASTInfixExpression left = (ASTInfixExpression) e.getLeftOperand();
-        assertTrue(JavaRuleUtil.isStringConcatExpr(left));
-
-        //                      This is (i+j)
-        //                  vvvvvvvvvvvvvvvvvvvvv
-        assertEquals(listOf(left.getLeftOperand(), left.getRightOperand(), e.getRightOperand()),
-                     JavaRuleUtil.flattenOperands(e).toList());
     }
 
 }
